chatbot-python/requirements.txt
fastapi==0.110.0
openai==1.12.0
supabase==2.15.2
tenacity==8.2.3
python-dotenv==1.0.1
uvicorn==0.29.0
anthropic==0.52.1 